+++ flint/convol.py
@@ -10,9 +10,11 @@ from pathlib import Path
 from typing import Collection, List, NamedTuple, Optional
 
 import astropy.units as u
+import numpy as np
+from astropy.io import fits
 from astropy.wcs import FITSFixedWarning
-from racs_tools import beamcon_2D
-from radio_beam import Beam
+from racs_tools import beamcon_2D, beamcon_3D
+from radio_beam import Beam, Beams
 
 from flint.logging import logger
 
@@ -51,6 +53,127 @@ class BeamShape(NamedTuple):
             bmin_arcsec=radio_beam.minor.to(u.arcsecond).value,  # type: ignore
             bpa_deg=radio_beam.pa.to(u.degree).value,  # type: ignore
         )
+
+
+def check_if_cube_fits(fits_file: Path) -> bool:
+    """Check to see whether the data component of a FITS images is a cube.
+    Returns ``True`` is the data-shape needs 3-dimensions to be represented.
+
+    Note: Unclear on usefulness
+
+    Args:
+        fits_file (Path): FITS file that will be examinined
+
+    Returns:
+        bool: Whether the input FITS file is a cube or not.
+    """
+
+    try:
+        squeeze_data = np.squeeze(fits.getdata(fits_file))  # type: ignore
+    except FileNotFoundError:
+        return False
+    except OSError:
+        return False
+
+    return len(squeeze_data.shape) == 3
+
+
+def get_cube_common_beam(
+    cube_paths: Collection[Path], cutoff: Optional[float] = None
+) -> List[BeamShape]:
+    """Given a set of input cube FITS files, compute a common beam
+    for each channel.
+
+    Args:
+        cube_paths (Collection[Path]): Set of cube FITS files to inspect to derive a common beam
+        cutoff (Optional[float], optional): A cutoff value, in arcsec, that specifies the maximum BMAJ allowed. Defaults to None.
+
+    Returns:
+        List[BeamShape]: List of target beam shapes to use, corresponding to each channel
+    """
+
+    _, common_beam_data_list = beamcon_3D.smooth_fits_cube(
+        infiles_list=list(cube_paths),
+        dryrun=True,
+        cutoff=cutoff,
+        mode="natural",
+        conv_mode="robust",
+        ncores=1,
+    )
+    # Make proper check here that accounts for NaNs
+    for file in common_beam_data_list:
+        assert all(
+            (file[0].major == common_beam_data_list[0][0].major)
+            | np.isnan(file[0].major)
+        )
+        assert all(
+            (file[0].minor == common_beam_data_list[0][0].minor)
+            | np.isnan(file[0].minor)
+        )
+        assert all(
+            (file[0].pa == common_beam_data_list[0][0].pa) | np.isnan(file[0].pa)
+        )
+
+    first_cube_fits_beam = common_beam_data_list[0][0]
+    assert isinstance(
+        first_cube_fits_beam, Beams
+    ), f"Unexpected type for common beams. Expected Beams, got {type(first_cube_fits_beam)}"
+
+    beam_shape_list = [
+        BeamShape.from_radio_beam(radio_beam=beam)  # type: ignore
+        for beam in first_cube_fits_beam
+    ]
+    return beam_shape_list
+
+
+def convolve_cubes(
+    cube_paths: Collection[Path],
+    beam_shapes: List[BeamShape],
+    cutoff: Optional[float] = None,
+    convol_suffix: str = "conv",
+) -> Collection[Path]:
+    logger.info(f"Will attempt to convol {len(cube_paths)} cubes")
+    if cutoff:
+        logger.info(f"Supplied cutoff {cutoff}")
+
+    # Extractubg the beam properties
+    beam_major_list = [float(beam.bmaj_arcsec) for beam in beam_shapes]
+    beam_minor_list = [float(beam.bmin_arcsec) for beam in beam_shapes]
+    beam_pa_list = [float(beam.bpa_deg) for beam in beam_shapes]
+
+    # Sanity test
+    assert len(beam_major_list) == len(beam_minor_list) == len(beam_pa_list)
+
+    logger.info("Convoling cubes")
+    cube_data_list, _ = beamcon_3D.smooth_fits_cube(
+        infiles_list=list(cube_paths),
+        dryrun=False,
+        cutoff=cutoff,
+        mode="natural",
+        conv_mode="robust",
+        bmaj=beam_major_list,
+        bmin=beam_minor_list,
+        bpa=beam_pa_list,
+        suffix=convol_suffix,
+    )
+
+    # Construct the name of the new file created. For the moment this is done
+    # manually as it is not part of the returned object
+    # TODO: Extend the return struct from beamcon_3D to include output name
+    convol_cubes_path = [
+        Path(cube_data.filename).with_suffix(f".{convol_suffix}.fits")
+        for cube_data in cube_data_list
+    ]
+
+    # Show the mapping as a sanity check
+    for input_cube, output_cube in zip(list(cube_paths), convol_cubes_path):
+        logger.info(f"{input_cube=} convolved to {output_cube}")
+
+    # Trust no one
+    assert all(
+        [p.exists() for p in convol_cubes_path]
+    ), "A convolved cube does not exist"
+    return convol_cubes_path
 
 
 def get_common_beam(
@@ -158,6 +281,12 @@ def get_parser() -> ArgumentParser:
         default="conv",
         help="The suffix added to convolved images. ",
     )
+    convol_parser.add_argument(
+        "--cubes",
+        action="store_true",
+        default=False,
+        help="Treat the input files as cubes and use the corresponding 3D beam selection and convolution. ",
+    )
 
     maxbeam_parser = subparsers.add_parser(
         "maxbeam", help="Find the optimal beam size for a set of images."
@@ -176,6 +305,23 @@ def get_parser() -> ArgumentParser:
         help="Beams whose major-axis are larger then this (in arcseconds) are ignored from the calculation of the optimal beam.",
     )
 
+    cubemaxbeams_parser = subparsers.add_parser(
+        "cubemaxbeam",
+        help="Calculate the set of common beams across channels in a set of cubes",
+    )
+    cubemaxbeams_parser.add_argument(
+        "cubes",
+        type=Path,
+        nargs="+",
+        help="The images that will be convolved to a common resolution",
+    )
+    cubemaxbeams_parser.add_argument(
+        "--cutoff",
+        type=float,
+        default=None,
+        help="Beams whose major-axis are larger then this (in arcseconds) are ignored from the calculation of the optimal beam.",
+    )
+
     return parser
 
 
@@ -187,13 +333,38 @@ def cli() -> None:
     if args.mode == "maxbeam":
         get_common_beam(image_paths=args.images, cutoff=args.cutoff)
     if args.mode == "convol":
-        common_beam = get_common_beam(image_paths=args.images, cutoff=args.cutoff)
-        _ = convolve_images(
-            image_paths=args.images,
-            beam_shape=common_beam,
-            cutoff=args.cutoff,
-            convol_suffix=args.convol_suffix,
+        if args.cubes:
+            assert all(
+                [check_if_cube_fits(fits_file=f) for f in args.images]
+            ), "Not all input files are FITS cubes"
+            common_beams = get_cube_common_beam(
+                cube_paths=args.images, cutoff=args.cutoff
+            )
+            for image in args.images:
+                logger.info(f"Convoling {image}")
+                _ = convolve_cubes(
+                    cube_paths=[image],
+                    beam_shapes=common_beams,
+                    cutoff=args.cutoff,
+                    convol_suffix=args.convol_suffix,
+                )
+
+        else:
+            assert not all(
+                [check_if_cube_fits(fits_file=f) for f in args.images]
+            ), "Not all input files are FITS images (not cubes)"
+            common_beam = get_common_beam(image_paths=args.images, cutoff=args.cutoff)
+            _ = convolve_images(
+                image_paths=args.images,
+                beam_shape=common_beam,
+                cutoff=args.cutoff,
+                convol_suffix=args.convol_suffix,
+            )
+    if args.mode == "cubemaxbeam":
+        common_beam_shape_list = get_cube_common_beam(
+            cube_paths=args.cubes, cutoff=args.cutoff
         )
+        logger.info(f"Extracted {common_beam_shape_list=}")
 
 
 if __name__ == "__main__":